+++ database/migrations/2021_03_13_024704_create_pekerjaan_selesai_table.php
@@ -15,8 +15,8 @@ class CreatePekerjaanSelesaiTable extends Migration
     {
         Schema::create('pekerjaan_selesai', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('ccr_id')->unique();
-            $table->unsignedInteger('ppc_id');
+            $table->unsignedBigInteger('ccr_id')->unique();
+            $table->unsignedBigInteger('ppc_id');
             $table->date('date_finish');
 
             $table->foreign('ccr_id')->references('id')->on('ccr')->onDelete('cascade')->onUpdate('cascade');
